backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
langchain
langchain-ollama
langchain-core
python-dotenv 